# src/internal/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: Person Service API
  version: 1.0.0
servers:
  - url: http://localhost:8080
security: []
paths:
  /health:
    get:
      tags: [Service]
      summary: Check service readiness
      operationId: health
      responses:
        "200":
          description: Service and database are available
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: ok
        "503":
          description: Database is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/persons:
    get:
      tags: [Persons]
      summary: Get all persons
      operationId: listPersons
      responses:
        "200":
          description: All persons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PersonResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags: [Persons]
      summary: Create a person
      operationId: createPerson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePersonRequest"
      responses:
        "201":
          description: Person created
          headers:
            Location:
              description: Path to the created person
              required: true
              schema:
                type: string
                example: /api/v1/persons/1
        "400":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/persons/{id}:
    parameters:
      - $ref: "#/components/parameters/PersonID"
    get:
      tags: [Persons]
      summary: Get a person by ID
      operationId: getPerson
      responses:
        "200":
          description: Requested person
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    patch:
      tags: [Persons]
      summary: Partially update a person
      operationId: updatePerson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePersonRequest"
      responses:
        "200":
          description: Updated person
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: [Persons]
      summary: Delete a person
      operationId: deletePerson
      responses:
        "204":
          description: Person deleted
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  parameters:
    PersonID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int32
        minimum: 1
  schemas:
    CreatePersonRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        age:
          type: integer
          format: int32
          minimum: 0
          nullable: true
        address:
          type: string
          nullable: true
        work:
          type: string
          nullable: true
    UpdatePersonRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
        age:
          type: integer
          format: int32
          minimum: 0
        address:
          type: string
        work:
          type: string
    PersonResponse:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        age:
          type: integer
          format: int32
          nullable: true
        address:
          type: string
          nullable: true
        work:
          type: string
          nullable: true
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ValidationErrorResponse:
      type: object
      required: [message, errors]
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
    NotFound:
      description: Person not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"